+++ src/components/layout.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import Container from '@mui/material/Container';
 import Box from '@mui/material/Box';
-import { Header } from './header';
-import { Footer } from './footer';
+import { Header } from './Header';
+import { Footer } from './Footer';
 
 export const Layout: React.FC<{ children: React.ReactNode }> = ({ children }) => (
     <Box display="flex" flexDirection="column" minHeight="100vh">
